fix(song/tag): use f32 to store duration

+++ src/models/songs.rs
@@ -6,13 +6,13 @@ use std::borrow::Cow;
 use time::OffsetDateTime;
 use uuid::Uuid;
 
-#[derive(Debug, Identifiable, Queryable, Selectable, Clone, PartialEq, Eq)]
+#[derive(Debug, Identifiable, Queryable, Selectable, Clone, PartialEq)]
 #[diesel(table_name = songs)]
 #[diesel(check_for_backend(diesel::pg::Pg))]
 pub struct Song {
     pub id: Uuid,
     pub title: String,
-    pub duration: i32,
+    pub duration: f32,
     pub album_id: Uuid,
     pub music_folder_id: Uuid,
     pub path: String,
@@ -26,7 +26,7 @@ pub struct Song {
 #[diesel(table_name = songs)]
 pub struct NewOrUpdateSong<'a> {
     pub title: Cow<'a, str>,
-    pub duration: i32,
+    pub duration: f32,
     pub album_id: Uuid,
     pub music_folder_id: Uuid,
     pub path: Option<Cow<'a, str>>,
